requirements.txt
pillow
Click
pytest